admin/requirements.txt
appdirs==1.4.3
Django==1.10.6
django-money==0.10.2
packaging==16.8
py-moneyed==0.6.0
pyparsing==2.2.0
six==1.10.0
